requirements.txt
-e .
aws_cdk.core
aws_cdk.aws_s3
aws_cdk.aws_ecs
aws_cdk.aws_ec2
aws_cdk.aws_logs
aws_cdk.aws_lambda
aws_cdk.aws_ecs_patterns